abstrakt endre mottaker med API håndtering
