+++ skribenten-web/frontend/src/routes/saksnummer_/$saksId/brevvelger/kladd.$brevId.tsx
@@ -1,24 +1,20 @@
 import { css } from "@emotion/react";
 import { ArrowRightIcon } from "@navikt/aksel-icons";
 import { BodyShort, Button, Heading, HStack, Loader, VStack } from "@navikt/ds-react";
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
+import { useQuery } from "@tanstack/react-query";
 import { createFileRoute, useNavigate } from "@tanstack/react-router";
-import type { AxiosError } from "axios";
-import { useState } from "react";
 
-import { delvisOppdaterBrev, fjernOverstyrtMottaker, hentAlleBrevForSak } from "~/api/sak-api-endpoints";
+import { hentAlleBrevForSak } from "~/api/sak-api-endpoints";
 import { ApiError } from "~/components/ApiError";
 import { Divider } from "~/components/Divider";
+import EndreMottakerMedOppsummeringOgApiHåndtering from "~/components/EndreMottakerMedApiHåndtering";
 import LetterTemplateTags from "~/components/LetterTemplateTags";
-import OppsummeringAvMottaker from "~/components/OppsummeringAvMottaker";
 import { SlettBrev } from "~/components/SlettBrev";
-import { mapEndreMottakerValueTilMottaker } from "~/types/AdresseUtils";
 import { type LetterMetadata } from "~/types/apiTypes";
-import type { BrevInfo, DelvisOppdaterBrevResponse, Mottaker } from "~/types/brev";
+import type { BrevInfo } from "~/types/brev";
 import { SPRAAK_ENUM_TO_TEXT } from "~/types/nameMappings";
 
 import Oppsummeringspar from "../kvittering/-components/Oppsummeringspar";
-import { EndreMottakerModal } from "./$templateId/-components/endreMottaker/EndreMottaker";
 
 export const Route = createFileRoute("/saksnummer/$saksId/brevvelger/kladd/$brevId")({
   loader: async ({ context: { queryClient, getSakContextQueryOptions } }) => {
@@ -62,33 +58,7 @@ const BrevIkkeFunnet = (props: { brevId: string }) => {
 };
 
 const Brevmal = (props: { saksId: string; brev: BrevInfo; letterMetadata: LetterMetadata }) => {
-  const queryClient = useQueryClient();
   const navigate = useNavigate({ from: Route.fullPath });
-  const [modalÅpen, setModalÅpen] = useState<boolean>(false);
-
-  const mottakerMutation = useMutation<DelvisOppdaterBrevResponse, AxiosError, Mottaker>({
-    mutationFn: (mottaker) =>
-      delvisOppdaterBrev({
-        saksId: props.saksId,
-        brevId: props.brev.id,
-        mottaker: mottaker,
-      }),
-    onSuccess: (response) => {
-      queryClient.setQueryData(hentAlleBrevForSak.queryKey(props.saksId), (currentBrevInfo: BrevInfo[]) =>
-        currentBrevInfo.map((brev) => (brev.id === props.brev.id ? response.info : brev)),
-      );
-      setModalÅpen(false);
-    },
-  });
-
-  const fjernMottakerMutation = useMutation<void, AxiosError>({
-    mutationFn: () => fjernOverstyrtMottaker({ saksId: props.saksId, brevId: props.brev.id }),
-    onSuccess: () => {
-      queryClient.setQueryData(hentAlleBrevForSak.queryKey(props.saksId), (currentBrevInfo: BrevInfo[]) =>
-        currentBrevInfo.map((brev) => (brev.id === props.brev.id ? { ...props.brev, mottaker: null } : brev)),
-      );
-    },
-  });
 
   return (
     <div
@@ -122,40 +92,7 @@ const Brevmal = (props: { saksId: string; brev: BrevInfo; letterMetadata: Letter
         <Divider />
 
         <VStack gap="8">
-          <VStack gap="2">
-            <VStack>
-              <OppsummeringAvMottaker mottaker={props.brev.mottaker} saksId={props.saksId} withTitle />
-
-              {modalÅpen && (
-                <EndreMottakerModal
-                  error={mottakerMutation.error}
-                  isPending={mottakerMutation.isPending}
-                  onBekreftNyMottaker={(mottaker) => {
-                    mottakerMutation.mutate(mapEndreMottakerValueTilMottaker(mottaker));
-                  }}
-                  onClose={() => setModalÅpen(false)}
-                  resetOnBekreftState={() => mottakerMutation.reset()}
-                  åpen={modalÅpen}
-                />
-              )}
-            </VStack>
-            <HStack>
-              <Button onClick={() => setModalÅpen(true)} size="small" type="button" variant="secondary">
-                Endre mottaker
-              </Button>
-              {props.brev.mottaker !== null && (
-                <Button
-                  loading={fjernMottakerMutation.isPending}
-                  onClick={() => fjernMottakerMutation.mutate()}
-                  size="small"
-                  type="button"
-                  variant="tertiary"
-                >
-                  Tilbakestill mottaker
-                </Button>
-              )}
-            </HStack>
-          </VStack>
+          <EndreMottakerMedOppsummeringOgApiHåndtering brev={props.brev} saksId={props.saksId} />
           <Oppsummeringspar
             boldedTitle
             tittel={"Avsenderenhet"}
